requirements.txt
aiosmtplib==2.0.2
annotated-types==0.5.0
anyio==3.7.1
base58==2.1.1
bcrypt==4.0.1
beanie==1.22.6
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
fastapi-mail==1.4.1
h11==0.14.0
idna==3.4
ipfshttpclient==0.7.0
Jinja2==3.1.2
lazy-model==0.2.0
MarkupSafe==2.1.3
motor==3.3.1
multiaddr==0.0.9
netaddr==0.9.0
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
varint==1.0.2
pillow==10.0.1